Create CalcBMI.PY
TASK FOR Calculating BMI, storing in a CSV file and displaying a pie chart:-
Create a program for a BMI calculator that takes the name, age, height, and weight of a person and
stores the data into a file. Allow the application to show weather the user is Underweight,
Overweight, Obese, or Normal.
Make sure you add validation rules.
Add a ‘Show Graph’ button that display a pie chart of the data stored, showing the percentage Of people under certain BMI categories.
Challenge:
Can you allow the chart to show on the same window? You may need to use another library.